doc: Add running the example with CAN interface
Updated the documentation to include instructions on running the example with
the CAN interface.

Provided example commands to start the server and client processes using
the CAN interface.

This update enhances the example documentation by including steps for testing
with the CAN interface, making it easier for users to run and validate the CAN
communication in their environment.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>