pycsp: Ensure default interface is set correctly when rtable is not used
Add default interface in `pycsp_zmqhub_init`, `pycsp_can_socketcan_init`, and
`pycsp_kiss_init` functions.

This change ensures that the default interface is correctly marked as default
when `rtable` is not used, addressing a bug where the initialization would fail
without `rtable`.

This fix is crucial for setups that do not utilize the routing table
(`CSP_USE_RTABLE`).

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>