python: bindings: Add Python binding for csp_get_buf_free
This commit introduces the `get_buf_free(node, timeout=1000)` function in the
Python bindings. It wraps the `csp_get_buf_free` function, allowing users to
query the number of free CSP buffers available on a remote node.

The function returns the value as a Python integer or raises a PyErr_Error if
the transaction fails.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>